back/requirements.txt
fastapi
uvicorn
numpy
opencv-python
Pillow
requests
python-multipart